Moved message daemon to the webserver's folder
This was a more fitting location, as the message daemon was not intended anymore to run on the RPi (client).
